refactor - Line 생성 메소드 Line 클래스로 이관
