+++ src/main/java/nextstep/ladder/domain/Line.java
@@ -1,13 +1,40 @@
 package nextstep.ladder.domain;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Line {
 
+    private static final int ZERO = 0;
+    private static final int ONE = 1;
+
     private List<Point> points;
 
-    public Line(List<Point> points) {
+    public Line() {}
+
+    private Line(List<Point> points) {
         this.points = points;
+    }
+
+    public Line initLine(Users users) {
+        List<Point> points = new ArrayList<>();
+        Direction direction = null;
+
+        int bound = users.getUsers().size();
+        for (int number = ZERO; number < bound; number++) {
+            int condition = minusIndex(number, bound);
+            direction = decideDirection(direction, condition, new RandomDirectionStrategy());
+            points.add(new Point(number, direction));
+        }
+
+        return new Line(points);
+    }
+
+    private int minusIndex(int number, int bound) {
+        if(bound - ONE - number == ZERO) {
+            return -ONE;
+        }
+        return number;
     }
 
     public List<Point> getPoints() {
